requirements.txt
PyYAML==5.2
praw==6.4.0
